blink_ai_server/requirements.txt
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
aiofiles==23.1.0
requests==2.31.0
# 量化相关依赖
onnx==1.14.1
onnxconverter-common==1.13.0
psutil==5.9.5
torch==2.0.1
torchvision==0.15.2
